everything is rearranged, popup works
