--- analytics.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-  <script async defer src="/public/bundle.js"></script>
-  <link rel="stylesheet" href="/public/css/page.css" />
-  <title>Degender Analytics</title>
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-</html>
