+++ analytics/ducks/index.jsx
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux'
+
+const rootReducer = combineReducers({
+  button: require('./button').default
+})
+
+export default rootReducer
