+++ analytics/main.jsx
@@ -1,25 +1,24 @@
 import React from 'react'
 import { render } from 'react-dom'
 import Container from './components/Container'
-import Popup from './components/Popup'
 
 import { Provider } from 'react-redux'
 import store from './store'
 
 //seems like react router doesn't work without a server :((((
 //import { Router, Route, IndexRedirect, browserHistory } from 'react-router'
-import * as firebase from 'firebase'
+// import * as firebase from 'firebase'
 
-const config = {
-    apiKey: "AIzaSyDhVC1K39EUPWBekzHXgltSe1xkQR6US-c",
-    authDomain: "degendertheinternet.firebaseapp.com",
-    databaseURL: "https://degendertheinternet.firebaseio.com",
-    storageBucket: "degendertheinternet.appspot.com",
-    messagingSenderId: "919863122992"
-};
+// const config = {
+//     apiKey: "AIzaSyDhVC1K39EUPWBekzHXgltSe1xkQR6US-c",
+//     authDomain: "degendertheinternet.firebaseapp.com",
+//     databaseURL: "https://degendertheinternet.firebaseio.com",
+//     storageBucket: "degendertheinternet.appspot.com",
+//     messagingSenderId: "919863122992"
+// };
 
 
-firebase.initializeApp(config);
+// firebase.initializeApp(config);
 
 if (typeof window !== 'undefined') {
     window.React = React;
@@ -27,8 +26,8 @@ if (typeof window !== 'undefined') {
 
 render (
         <Provider store={store}>
-          <Popup />
+          <Container />
         </Provider>
         ,
-  document.getElementById('app')
+  document.getElementById('analytics')
 )
