+++ popup/ducks/button.jsx
@@ -1,0 +1,11 @@
+//actions
+
+//reducer
+export default function reducer (state={}, action) {
+  switch (action.type) {
+    default:
+      return state
+  }
+}
+
+//action creators
