--- public/popup.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-
-  <link rel="stylesheet" href="/public/css/popup.css" />
-  </head>
-  <body>
-    <div id='popup-box'>
-      <button>Degender Current Page</button>
-      <ul>
-        <li>Settings</li>
-        <li>Show Altered Words</li>
-        <li><a href='/analytics.html' target='_blank'>Analyze this page</a></li>
-      </ul>
-    </div>
-  </body>
-</html>
